wms_project/requirements.txt
# Django Backend
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
gunicorn>=21.2

# Google Sheets Sync
gspread>=6.0
google-auth>=2.28
google-auth-oauthlib>=1.2

# Streamlit Frontend
streamlit>=1.32
requests>=2.31
pandas>=2.0
